pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ducklake-dataframe"
version = "0.2.0"
description = "DataFrame integration for DuckLake catalogs — Polars, pandas, and more"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Pedro Holanda", email = "pedroholanda@gmail.com" },
]
keywords = ["ducklake", "polars", "pandas", "data-lake", "parquet", "dataframe"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pyarrow>=10.0",
]

[project.urls]
Homepage = "https://github.com/pdet/ducklake-dataframe"
Repository = "https://github.com/pdet/ducklake-dataframe"
Issues = "https://github.com/pdet/ducklake-dataframe/issues"

[project.optional-dependencies]
polars = ["polars>=1.0"]
pandas = ["pandas>=1.5"]
duckdb = ["duckdb>=1.2"]
postgres = ["psycopg2-binary>=2.9"]
cloud = ["fsspec>=2023.1"]
s3 = ["fsspec>=2023.1", "s3fs"]
gcs = ["fsspec>=2023.1", "gcsfs"]
azure = ["fsspec>=2023.1", "adlfs"]
all = [
    "polars>=1.0",
    "pandas>=1.5",
    "duckdb>=1.2",
    "psycopg2-binary>=2.9",
    "fsspec>=2023.1",
    "s3fs",
    "gcsfs",
    "adlfs",
]
dev = [
    "pytest>=7.0",
    "pytest-xdist",
    "pytest-timeout",
    "duckdb>=1.2",
    "polars>=1.0",
    "pandas>=1.5",
    "hypothesis>=6.0",
    "pytz",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ducklake_core", "src/ducklake_polars", "src/ducklake_pandas"]

[tool.pytest.ini_options]
testpaths = ["tests", "tests_pandas"]
markers = [
    "postgres: PostgreSQL integration tests (require DUCKLAKE_PG_DSN env var)",
]
